// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  INACTIVE: 'INACTIVE'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const SemesterType = {
  ODD: 'ODD',
  EVEN: 'EVEN'
} as const

export type SemesterType = (typeof SemesterType)[keyof typeof SemesterType]


export const AcademicPeriodStatus = {
  PLANNED: 'PLANNED',
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type AcademicPeriodStatus = (typeof AcademicPeriodStatus)[keyof typeof AcademicPeriodStatus]


export const CourseOfferingType = {
  REGULAR: 'REGULAR',
  EXTRAORDINARY: 'EXTRAORDINARY'
} as const

export type CourseOfferingType = (typeof CourseOfferingType)[keyof typeof CourseOfferingType]


export const CourseOfferingStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type CourseOfferingStatus = (typeof CourseOfferingStatus)[keyof typeof CourseOfferingStatus]


export const GroupStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus]


export const EnrollmentStatus = {
  IN_PROGRESS: 'IN_PROGRESS',
  PASSED: 'PASSED',
  FAILED: 'FAILED',
  WITHDRAWN: 'WITHDRAWN',
  CANCELLED: 'CANCELLED'
} as const

export type EnrollmentStatus = (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus]


export const EnrollmentResult = {
  PASSED: 'PASSED',
  FAILED: 'FAILED'
} as const

export type EnrollmentResult = (typeof EnrollmentResult)[keyof typeof EnrollmentResult]


export const FinancialObligationType = {
  SEMESTER: 'SEMESTER',
  MONTHLY: 'MONTHLY',
  COURSE_RECOVERY: 'COURSE_RECOVERY',
  OTHER: 'OTHER'
} as const

export type FinancialObligationType = (typeof FinancialObligationType)[keyof typeof FinancialObligationType]


export const FinancialObligationStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  OVERDUE: 'OVERDUE',
  CANCELLED: 'CANCELLED'
} as const

export type FinancialObligationStatus = (typeof FinancialObligationStatus)[keyof typeof FinancialObligationStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  TRANSFER: 'TRANSFER',
  GATEWAY: 'GATEWAY'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]